\name{rsem-package}
\alias{rsem-package}
\docType{package}
\title{Robust Structural Equation Modeling with Missing Data and Auxiliary Variables
}
\description{
A robust procedure is implemented to estimate means and covariance matrix of multiple variables with missing data using Huber weight and then to estimate a structural equation model.
}
\details{
\tabular{ll}{
Package: \tab rsem\cr
Type: \tab Package\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{
Ke-Hai Yuan and Zhiyong Zhang
Maintainer: Zhiyong Zhang <zhiyongzhang@nd.edu>
}
\references{
Yuan, K.-H., & Zhang, Z. (2012). Robust structural equation modeling with missing data and auxiliary variables. Psychometrika, 77(4), 803-826. https://doi.org/10.1007/s11336-012-9282-4
}

