\name{rsem.DP}
\alias{rsem.DP}
\title{
Generate a duplication matrix
}
\description{
Generate a duplication matrix
}
\usage{
rsem.DP(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A matrix
}
}

\references{
Ke-Hai Yuan and Zhiyong Zhang (2011) Robust Structural Equation Modeling with Missing Data and Auxiliary Variables
}
\author{
Ke-Hai Yuan and Zhiyong Zhang
}

\examples{
x<-array(1:6, c(2,3))
rsem.DP(x)
}

