\name{rsem.gname}
\alias{rsem.gname}
\title{
Internal function
}
\description{
Internal function
}
\usage{
rsem.gname(name)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{name}{
Variable names.
}
}

\references{
Ke-Hai Yuan and Zhiyong Zhang (2011) Robust Structural Equation Modeling with Missing Data and Auxiliary Variables
}
\author{
Ke-Hai Yuan and Zhiyong Zhang
}


