\name{rsem.switch.gamma}
\alias{rsem.switch.gamma}
\title{
Internal function
}
\description{
Internal function
}
\usage{
rsem.switch.gamma(gamma, ov.names)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{gamma}{
Robust covariance matrix for saturated mean and covariances
}

  \item{ov.names}{
Observed variable names.
}
}

\references{
Ke-Hai Yuan and Zhiyong Zhang (2011) Robust Structural Equation Modeling with Missing Data and Auxiliary Variables
}
\author{
Ke-Hai Yuan and Zhiyong Zhang
}


