\name{rsem.vec}
\alias{rsem.vec}
\title{
Stacking a matrix to a vector
}
\description{
Stacking a matrix to a vector
}
\usage{
rsem.vec(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A matrix
}
}

\references{
Ke-Hai Yuan and Zhiyong Zhang (2011) Robust Structural Equation Modeling with Missing Data and Auxiliary Variables
}
\author{
Ke-Hai Yuan and Zhiyong Zhang
}

\examples{
x<-array(1:6, c(2,3))
rsem.vec(x)
}

