\name{rsem-package}
\alias{rsem-package}
\docType{package}
\title{Robust Structural Equation Modeling with Missing Data and Auxiliary
}
\description{
This package estimates means and covariance matrix of multiple variables with missing data using Huber weight and then estimates a SEM model using either lavaan or EQS.
}
\details{
\tabular{ll}{
Package: \tab rsem\cr
Type: \tab Package\cr
Version: \tab 0.4.3\cr
Date: \tab 2010-12-27\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{
Ke-Hai Yuan and Zhiyong Zhang
Maintainer: Zhiyong Zhang <zhiyongzhang@nd.edu>
}
\references{
Ke-Hai Yuan and Zhiyong Zhang (2011) Robust Structural Equation Modeling with Missing Data and Auxiliary Variables
}

